Update CrafterName on items when saving
Adds a pass to the save profile code that re-sets the crafterName field on any item that was made by the current player. This ensures if their name was changed the item shows as crafted by their new name. This is done by just matching the crafter ID values to the current profile player ID.