Load Shared Item Data at runtime from CSV file
Instead of having the Shared Item Data db hard coded in c# and built in at compile time, read it from external file at runtime. File format is CSV and the current version of the hard coded database is converted and added as a CSV file.

If the CSV file cannot be found or read during execution, the program won't crash but all items will be anonymous. Current solution will indicate the issue in the debug console but not to the user. The latter can be added later, this commit is just to go from internal C# to external CSV file.