Update Item.cs
This will at least prevent Loki from crashing but maybe there are better ways of solving this.

Another thought: I'd like to insert the actual internal game id (ie name) of the item in the inventory slot, instead of the fallback value 'item_name'. That way you can see what the name is that is missing, to be added in the SharedItemData DB. WPF is not my expertise so I don't know how to make a fallback that use the name of an item. Well, you could replace "item_data" with the item.name in code but that seems backwards...