Fix crash caused by bug in .NET 7
This prevented characters being loaded in .NET 7 due to a new bug where if you create a new colour from scalar values, the ToString method on the class `Color` throws a FormatException.
Need to look into that later and report it with a minimal example and/or just fix it and give MS a pull request.